Some work done with the icons to get working.
Experimented with font icons vs SVG to get what I needed. Still not entire sure if current setup with SVG images is the best way to go but but for now that's the way. Font icons was giving issues with rollover not working accurately. SVG inline while flexible with changing colors involved imbedded lots of XML code for each instance of a particular icon.