Fixed issues with browser zoom and broken layout on iOS Safari.
When zooming in the browser the projects layout breaks dramatically. Specifically the image previews with double or triple layouts. While not entirely fixed the problem only occurs at much higher zoom levels (150% or greater with a narrow browser window)

The layout issues on iOS Safari seems to be fixed entirely though.